Rename cpp file
